Moved module for command checks
